requirements.txt
beautifulsoup4==4.13.3
cbor2==5.6.5
certifi==2025.1.31
charset-normalizer==3.4.1
frozendict==2.4.6
idna==3.10
lgpio==0.2.2.0
luma.core==2.4.2
luma.oled==3.14.0
multitasking==0.0.11
numpy==2.2.4
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
pyftdi==0.56.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyusb==1.3.1
requests==2.32.3
rpi-lgpio==0.6
RPi.GPIO==0.7.1
six==1.17.0
smbus2==0.5.0
soupsieve==2.6
spidev==3.6
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yfinance==0.2.55
